Logistics: fix several more guaranteed Dapper crashes found via exhaustive audit
Went through every remaining QueryAsync<long>/QuerySingleAsync<long>/
tuple materialization in Helpers/Logistics against the live schema.
Found and fixed:
- PerformanceSnapshotHelper.StartRefreshAsync: dbo.RefreshLog.RunId is
  int, not long — every scheduled/manual refresh (Full Refresh,
  TurnsValClass Refresh, MRP History Refresh) was throwing here before
  any sync work ever ran.
- PerformanceSnapshotHelper.UpsertMaterialReceiptHistoryAsync: the
  log.Vendor lookup tuple declared VendorId as long against the real
  int column.
- OtifMetricRawRow (log.DailyPerformance): OtifOnTimeCount/
  OtifTotalCount are int, not long? — broke the Management dashboard's
  /otif-metrics route on every call.
- ConsignmentTrackerHelper.AddManualDeliveryAsync/CreateDeclarationAsync:
  log.ConsignmentDelivery.DeliveryId and log.ConsignmentDeclaration.
  DeclarationId are both int, not long.
- InboundShipmentHelper.RemoveManualItemAsync/CancelShipmentAsync:
  log.ManualInboundItem.ItemId and log.PurchaseOrderSuggestion.
  SuggestionId are both int, not long.

Same established bug class as the earlier fixes this session — Dapper's
strict positional-record/tuple materialization throws based on the
query's column schema alone, regardless of row content, whenever a
declared long/long? doesn't exactly match a real int column.

+++ dotnet/NormantonNexus/Helpers/Logistics/ConsignmentTrackerHelper.cs
@@ -173,7 +173,8 @@ internal static class ConsignmentTrackerHelper
     internal static async Task<long> AddManualDeliveryAsync(INexusOperationsDb db, long vendorId, AddManualConsignmentDeliveryRequest body, string? username, CancellationToken ct)
     {
         using var connection = await db.CreateConnectionAsync(ct);
-        return await connection.QuerySingleAsync<long>(new CommandDefinition("""
+        // DeliveryId is int, matching log.ConsignmentDelivery.DeliveryId's real column type.
+        return await connection.QuerySingleAsync<int>(new CommandDefinition("""
             INSERT INTO log.ConsignmentDelivery
                 (VendorId, Material, MaterialDocument, MaterialDocItem, Quantity, Uom,
                  Container, BillOfLading, InvoiceNumber, DocumentDate, PostingDate, ExpiryDate,
@@ -590,7 +591,8 @@ internal static class ConsignmentTrackerHelper
         try
         {
             var totalQty = body.Lines.Sum(l => l.QtyAllocated);
-            var declarationId = await connection.QuerySingleAsync<long>(new CommandDefinition("""
+            // DeclarationId is int, matching log.ConsignmentDeclaration.DeclarationId's real column type.
+            var declarationId = await connection.QuerySingleAsync<int>(new CommandDefinition("""
                 INSERT INTO log.ConsignmentDeclaration (VendorId, Status, AllocationMethod, TotalQty, CreatedByUsername)
                 OUTPUT INSERTED.DeclarationId
                 VALUES (@vendorId, 'Draft', @allocationMethod, @totalQty, @username)
@@ -730,7 +732,8 @@ internal static class ConsignmentTrackerHelper
             if (status is null) throw new NexusNotFoundException("Declaration not found.");
             if (status != "Draft") throw new NexusValidationException($"Declaration is already {status}, not Draft.");
 
-            var lines = await connection.QueryAsync<(long DeliveryId, decimal QtyAllocated)>(new CommandDefinition(
+            // DeliveryId is int, matching log.ConsignmentDeclarationLine.DeliveryId's real column type.
+            var lines = await connection.QueryAsync<(int DeliveryId, decimal QtyAllocated)>(new CommandDefinition(
                 "SELECT DeliveryId, QtyAllocated FROM log.ConsignmentDeclarationLine WHERE DeclarationId = @declarationId",
                 new { declarationId }, transaction, cancellationToken: ct));
